% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_spp_model}
\alias{get_spp_model}
\title{Get Species Detection Model}
\usage{
get_spp_model()
}
\arguments{
\item{reload}{Logical. If TRUE, forces reloading of the model from file.
Default is FALSE.}
}
\value{
A fitted GLM object used for species detection predictions.
}
\description{
Loads the fitted GLM model for species detection from the package data.
The model is cached after first load for improved performance.
}
\details{
The model file should be located in the package's extdata directory
as 'model.rda'. The function will cache the model in memory after first
load to improve performance on subsequent calls.
}
\examples{
\dontrun{
model <- get_spp_model()
}

}
