% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mod.R
\name{plot_model}
\alias{plot_model}
\title{Plot model fit with confidence intervals and detection zones}
\usage{
plot_model(reads, labels = NULL, model_name = "DLOOP")
}
\arguments{
\item{reads}{Numeric vector of read counts}

\item{labels}{Character string vector of the label names}

\item{model_name}{Model to use - either DORY or DLOOP model}
}
\value{
ggplot2 object
}
\description{
Plot model fit with confidence intervals and detection zones
}
