% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mod.R
\name{plot_modfit}
\alias{plot_modfit}
\title{Plot model fit with confidence intervals and detection zones}
\usage{
plot_modfit(reads_vec, mod)
}
\arguments{
\item{reads_vec}{Numeric vector of read counts}

\item{mod}{Model object for predictions}
}
\value{
ggplot2 object
}
\description{
Plot model fit with confidence intervals and detection zones
}
