% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mod.R
\name{plot_model}
\alias{plot_model}
\title{Plot model fit with confidence intervals and detection zones}
\usage{
plot_model(reads_vec, model_name = c("DORY", "DLOOP"))
}
\arguments{
\item{reads_vec}{Numeric vector of read counts}

\item{model_name}{Model to use - either DORY or DLOOP model}
}
\value{
ggplot2 object
}
\description{
Plot model fit with confidence intervals and detection zones
}
